CouchDomainDocTypeChangeProvider
